Add files via upload
Once entering user name and password in the login page, you will be directed to order list, which has multiple pages of orders. Click the order No., and you'll be shown the details of the order, including a form of the names and DNA sequences within this order. Sensitive data like user name and primer details have been edited.